requirements.txt
python-telegram-bot
PyYAML
flask
rq
rq-scheduler
flask-sqlalchemy
Flask-HTTPAuth
requests
retrying
